services/neoantigen-pipeline/neoantigen-predictor/requirements.txt
boto3>=1.34,<2
nats-py>=2.7,<3
mhcflurry>=2.0,<3
numpy>=1.26,<2
pandas>=2.1,<3
